+++ src/ace_time/BasicZoneProcessor.h
@@ -9,7 +9,6 @@
 #include <string.h> // strchr()
 #include <stdint.h>
 #include <AceCommon.h> // copyReplaceChar()
-#include "../zoneinfo/ZonePolicy.h"
 #include "../zoneinfo/ZoneInfo.h"
 #include "../zoneinfo/Brokers.h"
 #include "common/common.h" // kAbbrevSize
@@ -474,8 +473,8 @@ class BasicZoneProcessorTemplate: public ZoneProcessor {
       mEpochYear = Epoch::currentEpochYear();
       mNumTransitions = 0; // clear cache
 
-      if (year < mZoneInfoBroker.zoneContext()->startYear - 1
-          || mZoneInfoBroker.zoneContext()->untilYear < year) {
+      if (year < mZoneInfoBroker.zoneContext().startYear() - 1
+          || mZoneInfoBroker.zoneContext().untilYear() < year) {
         return false;
       }
 
@@ -865,9 +864,9 @@ class BasicZoneProcessorTemplate: public ZoneProcessor {
      */
     static int16_t calcRuleOffsetMinutes(int16_t prevEffectiveOffsetMinutes,
         int16_t currentBaseOffsetMinutes, uint8_t atSuffix) {
-      if (atSuffix == internal::ZoneContext::kSuffixW) {
+      if (atSuffix == basic::ZoneContext::kSuffixW) {
         return prevEffectiveOffsetMinutes;
-      } else if (atSuffix == internal::ZoneContext::kSuffixS) {
+      } else if (atSuffix == basic::ZoneContext::kSuffixS) {
         return currentBaseOffsetMinutes;
       } else { // 'u', 'g' or 'z'
         return 0;