+++ src/tzonedb/zone_infos.h
@@ -86,7 +86,7 @@ namespace tzonedb {
 extern const char kTzDatabaseVersion[];
 
 // Metadata about the zonedb files.
-extern const internal::ZoneContext kZoneContext;
+extern const basic::ZoneContext kZoneContext;
 
 //---------------------------------------------------------------------------
 // Supported zones: 11