+++ src/tzonedbx/zone_policies.h
@@ -73,7 +73,7 @@
 #ifndef ACE_TIME_TZONEDBX_ZONE_POLICIES_H
 #define ACE_TIME_TZONEDBX_ZONE_POLICIES_H
 
-#include <zoneinfo/ZonePolicy.h>
+#include <zoneinfo/ZoneInfo.h>
 
 namespace ace_time {
 namespace tzonedbx {